Can't have isa() check on parentage, make cast_as throws an exception on a null pointer.
Damn, type_info::before() didn't do what I thought it was.
For some reason dynamic_cast<>() isn't throwing an exception when it should, so throw one ourselves when it returns a null pointer.
In 'parse', just try to cast to the NMEA/binary types and do nothing on the exception.